+++ src/Selenium2Library/keywords/_cookie.py
@@ -1,4 +1,4 @@
-from keywordgroup import KeywordGroup
+from .keywordgroup import KeywordGroup
 
 class _CookieKeywords(KeywordGroup):