+++ src/Selenium2Library/keywords/_formelement.py
@@ -1,5 +1,5 @@
 import os
-from keywordgroup import KeywordGroup
+from .keywordgroup import KeywordGroup
 
 class _FormElementKeywords(KeywordGroup):