+++ src/Selenium2Library/keywords/_logging.py
@@ -1,8 +1,7 @@
 import os
-import sys
 from robot.variables import GLOBAL_VARIABLES
 from robot.api import logger
-from keywordgroup import KeywordGroup
+from .keywordgroup import KeywordGroup
 
 class _LoggingKeywords(KeywordGroup):