+++ src/Selenium2Library/locators/__init__.py
@@ -1,6 +1,6 @@
-from elementfinder import ElementFinder
-from tableelementfinder import TableElementFinder
-from windowmanager import WindowManager
+from .elementfinder import ElementFinder
+from .tableelementfinder import TableElementFinder
+from .windowmanager import WindowManager
 
 __all__ = [
     "ElementFinder",