+++ src/Selenium2Library/keywords/_javascript.py
@@ -86,9 +86,11 @@ class _JavaScriptKeywords(KeywordGroup):
         JavaScript. Return values are converted to the appropriate type in
         Python, including WebElements.
 
-        Example:
-        | Execute JavaScript | window.my_js_function('arg1', 'arg2') |
-        | Execute JavaScript | ${CURDIR}/js_to_execute.js |
+        Examples:
+        | Execute JavaScript | window.my_js_function('arg1', 'arg2') |               |
+        | Execute JavaScript | ${CURDIR}/js_to_execute.js            |               |
+        | ${sum}=            | Execute JavaScript                    | return 1 + 1; |
+        | Should Be Equal    | ${sum}                                | ${2}          |
         """
         js = self._get_javascript_to_execute(''.join(code))
         self._info("Executing JavaScript:\n%s" % js)
@@ -105,9 +107,14 @@ class _JavaScriptKeywords(KeywordGroup):
         Scripts must complete within the script timeout or this keyword will
         fail. See the `Timeouts` section for more information.
 
-        Example:
+        Examples:
         | Execute Async JavaScript | var callback = arguments[arguments.length - 1]; | window.setTimeout(callback, 2000); |
-        | Execute Async JavaScript | ${CURDIR}/async_js_to_execute.js | |
+        | Execute Async JavaScript | ${CURDIR}/async_js_to_execute.js                |                                    |
+        | ${retval}=               | Execute Async JavaScript                        |                                    |
+        | ...                      | var callback = arguments[arguments.length - 1]; |                                    |
+        | ...                      | function answer(){callback("text");};           |                                    |
+        | ...                      | window.setTimeout(answer, 2000);                |                                    |
+        | Should Be Equal          | ${retval}                                       | text                               |
         """
         js = self._get_javascript_to_execute(''.join(code))
         self._info("Executing Asynchronous JavaScript:\n%s" % js)