+++ CHANGES.rst
@@ -1,4 +1,4 @@
-Release Notes
+﻿Release Notes
 =============
 
 1.6 (unreleased)
@@ -10,9 +10,32 @@ Release Notes
   did not complete.
   [Mika Batsman][elizaleong][emanlove] 
 
-- Added support for negative indicies for rows and columns in table-related
+- Added support for negative indices for rows and columns in table-related
   keywords.
   [eweitz]
+
+- Added strategy for locating elements by partial link text with locator
+  prefix 'partial link'.
+  [lina1]
+
+- Added new keyword 'Clear Element Text' for clearing the text of text entry 
+  elements.
+  [emanlove]
+
+- Fixed issue where 'Select Window’ with url strategy fails to locate window
+  [laulaz]
+
+- Fixed issue where a non-string assigned to window.id caused
+  'Select Window' and 'Get Window *' keywords to fail.
+  [ombre42]
+
+- Allow using key attributes (default strategy) when the locator contains
+  a '=' by using the prefix 'default='. Also make locator prefixes
+  space-insensitive.
+  [ombre42]
+
+A big thank you to [eweitz] and [HelioGuilherme66] for getting the
+continuous integration builds to go green by fixing internal tests.
 
 1.5
 ---