+++ CHANGES.rst
@@ -18,6 +18,9 @@ Release Notes
   prefix 'partial link'.
   [lina1]
 
+- Fixed issue where 'Select Window’ with url strategy fails to locate window
+  [laulaz]
+
 1.5
 ---
 - Copy Desired Capabilities before modifying to prevent affecting future