+++ CHANGES.rst
@@ -1,8 +1,8 @@
 Release Notes
 =============
 
-1.6 (unreleased)
-----------------
+1.6
+---
 - Added examples to 'Execute Javascript' and 'Execute Async Javascript'
   keyword documentation.
   [ombre42]